fastapi-app/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
psycopg2-binary
